[add] Segmentation HomeScreen.dart
