+++ lib/main.dart
@@ -1,6 +1,7 @@
+import 'package:aiml_mobile_2024/screens/JoinMember.dart';
+import 'package:aiml_mobile_2024/screens/infra_info.dart';
 import 'package:flutter/material.dart';
-import '../widget/CommonScaffold.dart'; // CommonScaffold.dart 파일 import
-import '../screens/JoinMember.dart'; // JoinMember.dart 파일 import
+import '../screens/HomeScreen.dart';
 import '../screens/ShowCustomSearchMessage.dart'; // ShowCustomSearchMessages.dart 파일 import
 
 void main() {
@@ -16,21 +17,13 @@ class MyApp extends StatelessWidget {
       initialRoute: '/',
       routes: {
         '/': (context) => HomeScreen(),
+        // 이건 임시로 해놓은 화면
+        '/infra_info': (context) => JoinMember(),
+        // '/navigator': (context) => Navigator(),
+        // '/search_missing_person': (context) => SearchMissingPerson(),
         '/message': (context) => ShowCustomSearchMessage(),
       },
     );
   }
 }
 
-class HomeScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return const CommonScaffold(
-      title: Text('Home'), // 홈 화면의 타이틀
-      body: Center(
-        child: Text("이제 여기에다가 구현하시면 됩니다."
-            "다른 것들도 페이지 만들어서 비슷하게 위에다가 연결하시면 될 듯해요..."),
-        ),
-    );
-  }
-}
